feat: add comprehensive xUnit test suite (111 tests)
- Add EventManagement.Tests project with unit and integration tests
- Unit tests: User loyalty tier/discount logic, AuthService (register,
  login, change password)
- Integration tests: Auth, Events, Bookings, Reviews, Subscriptions,
  Tags, Categories, Dev endpoints using WebApplicationFactory backed by
  an isolated in-memory SQLite database per test
- Each test gets its own factory instance for full isolation
- Fix decimal ORDER BY in EventsController (SQLite requires CAST to REAL)
- Add public partial class Program required by WebApplicationFactory

+++ backend/EventManagement/Controllers/EventsController.cs
@@ -68,7 +68,7 @@ public class EventsController(AppDbContext db) : ControllerBase
         query = sortBy?.ToLower() switch
         {
             "popularity" => query.OrderByDescending(e => e.Bookings.Count(b => b.Status == StatusConfirmed)),
-            "price"      => query.OrderBy(e => e.Price),
+            "price"      => query.OrderBy(e => (double)e.Price),
             _            => query.OrderBy(e => e.StartDate)
         };
